+++ cli/src/semgrep/__init__.py
@@ -1,2 +1,2 @@
-__VERSION__ = "1.11.5"
+__VERSION__ = "1.12.1"
 __SEMGREP_VERSION__ = "1.100.0"